+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 ### Removed:
   * Deprecated IDL attribute `@Swift(ObjC)` was removed.
+  * Removed "-mergemanifest" command line option. Android manifest merger should be performed with external tools
+    instead (e.g. using Gradle).
 
 ## 8.8.0
 Release date: 2021-01-27